requirements.txt
django-environ==0.12.0
django-filter==25.1
drf-spectacular==0.28.0
factory_boy==3.3.3
ipython==9.2.0
pip==24.0
psycopg==3.2.9
psycopg2-binary==2.9.10
pytest-cov==6.1.1
pytest-django==4.11.1
python-decouple==3.8
